Time: 8 ms (51.53%), Space: 47.6 MB (74.31%) - LeetHub

+++ 2938-separate-black-and-white-balls/2938-separate-black-and-white-balls.java
@@ -8,7 +8,7 @@ class Solution {
 
       for (int i = 0; i < s.length(); i++) {
         if (s.charAt(i) == '0')
-          swap += (long) black;
+          swap += black;
         else
           ++black;
       }
